+++ tests/test_checks_base.py
@@ -4,7 +4,6 @@ from tests import factories
 from threedi_modelchecker.checks import geo_query
 from threedi_modelchecker.checks.base import _sqlalchemy_to_sqlite_types
 from threedi_modelchecker.checks.base import EnumCheck
-from threedi_modelchecker.checks.base import FileExistsCheck
 from threedi_modelchecker.checks.base import ForeignKeyCheck
 from threedi_modelchecker.checks.base import GeometryCheck
 from threedi_modelchecker.checks.base import GeometryTypeCheck
@@ -660,63 +659,6 @@ def test_length_geom_linestring_missing_epsg_from_global_settings(session):
     assert len(errors) == 1
 
 
-def test_file_exists_check_filesystem_ok(session, tmp_path):
-    (tmp_path / "somefile").touch()
-    factories.GlobalSettingsFactory(dem_file="somefile")
-    session.model_checker_context.base_path = tmp_path
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 0
-
-
-def test_file_exists_check_filesystem_ignore(session, tmp_path):
-    factories.GlobalSettingsFactory(dem_file="")
-    session.model_checker_context.base_path = tmp_path
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 0
-
-
-def test_file_exists_check_filesystem_err(session, tmp_path):
-    factories.GlobalSettingsFactory(dem_file="some/file")
-    session.model_checker_context.base_path = tmp_path
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 1
-
-
-def test_file_exists_check_context_ok(session):
-    factories.GlobalSettingsFactory(dem_file="some/file")
-    session.model_checker_context.available_rasters = {"dem_file"}
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 0
-
-
-def test_file_exists_check_context_err(session):
-    factories.GlobalSettingsFactory(dem_file="some/file")
-    session.model_checker_context.available_rasters = {"frict_coef_file"}
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 1
-
-
-def test_file_exists_check_context_ignore(session):
-    factories.GlobalSettingsFactory(dem_file="")
-    session.model_checker_context.available_rasters = {"frict_coef_file"}
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 0
-
-
-def test_file_exists_check_skip(session):
-    # no context, no check, no invalid records
-    factories.GlobalSettingsFactory(dem_file="some/file")
-    check = FileExistsCheck(column=models.GlobalSetting.dem_file)
-    invalid_rows = check.get_invalid(session)
-    assert len(invalid_rows) == 0
-
-
 @pytest.mark.parametrize(
     "min_value,max_value,left_inclusive,right_inclusive",
     [
@@ -744,16 +686,16 @@ def test_range_check_valid(
 
 
 @pytest.mark.parametrize(
-    "min_value,max_value,left_inclusive,right_inclusive",
+    "min_value,max_value,left_inclusive,right_inclusive,msg",
     [
-        (0, 42, True, False),
-        (42, 100, False, True),
-        (None, 42, True, False),
-        (42, None, False, False),
+        (0, 42, True, False, "{} is <0 and/or >=42"),
+        (42, 100, False, True, "{} is <=42 and/or >100"),
+        (None, 42, True, False, "{} is >=42"),
+        (42, None, False, False, "{} is <=42"),
     ],
 )
 def test_range_check_invalid(
-    session, min_value, max_value, left_inclusive, right_inclusive
+    session, min_value, max_value, left_inclusive, right_inclusive, msg
 ):
     factories.ConnectionNodeFactory(storage_area=42)
 
@@ -766,3 +708,5 @@ def test_range_check_invalid(
     )
     invalid_rows = check.get_invalid(session)
     assert len(invalid_rows) == 1
+
+    assert check.description() == msg.format("v2_connection_nodes.storage_area")